Fix deadlock if the cross slice SharedScan consumer is on slice 0 (#623)
When the consumer of the cross slice SharedScan is in slice 0 and the producer
is in another slice, a deadlock may occur. The deadlock could happen because QD
does not complete slice 0 until other slices will be completed. But the
producer's slice cannot finish until it receives notification from the consumer
about the completion of reading. The consumer does not send this notification
before all QEs are completed.

This patch fixes the deadlock by creating a list of cross-slice consumers for
the current slice and sending notifications before waiting for QEs completion.
We can omit the share_lk_ctxt check for NULL since only cross slice Shared Scans
are added to the list.

Original patch was modified, because notifying producers was rewriting by
commit d21b113d131b502c89b09486734cb3de21d08530. But the main idea of the
original was not changed.

(cherry picked from commit f6f95b6622a3fbdf441ec86164d5921b472694d7)

Co-authored-by: Evgeniy Ratkov <e.ratkov@arenadata.io>