Fix CTEs with a volatile target list or having qual for the cases of SegmentGeneral and General locuses
The paths with the General and SegmentGeneral locus types should return the same
result on each of the segments. But a volatile function does not guarantee the
same result for different calls. If the target list or having qual of a path
node with the General or SegmentGeneral locus type contains a volatile function,
then the locus of such a path node should be brought to SingleQE. In the current
implementation, path nodes with volatile functions in the target list or having
qual are brought to SingleQE not for all subqueries. If a volatile function is
in the target list or having qual of CTE, multiset functions,
union/except/intersect then they can hide the volatile function in the
underlying nodes, and the current implementation can't detect it. The volatile
function can be executed on more than one segment, and each of them will work
with a different result set, which is not correct.

This patch fixes the issue by add the processing of volatile functions in the
target list and having qual for path nodes with the General and SegmentGeneral
locus type to rest of subquery planning cases.

This is a forward port of 120aaef599cf4cce1fac430b833a1b74b4a5d8bd from 6X.